instr api interface
