+++ edc-tests/runtime/edc-mock/build.gradle.kts
@@ -23,6 +23,7 @@ plugins {
     `java-library`
     id("application")
     id("com.github.johnrengelman.shadow") version "8.1.1"
+    id("io.swagger.core.v3.swagger-gradle-plugin")
 }
